quote関数を25行以内にした
